Add repositories

+++ RockyShop/Controllers/ProductController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using RockyShop.DataAccess.Data;
+using RockyShop.DataAccess.Repository.Interfaces;
+using RockyShop.Model.Models;
 using RockyShop.Model.ViewModels;
 using RockyShop.Utility.Services;
 using RockyShop.Utility.Utilities;
@@ -12,21 +12,26 @@ namespace RockyShop.Controllers
     [Authorize(Roles = Constants.AdminRole)]
 	public class ProductController : Controller
 	{
-		private readonly AppDbContext _dbContext;
+		private readonly IProductRepository _productRepo;
+        private readonly IRepository<Category> _categoryRepo;
+        private readonly IRepository<ApplicationType> _appTypeRepo;
         private readonly ProductImageService _productImageService;
 
-        public ProductController(AppDbContext dbContext, ProductImageService productImageService)
+        public ProductController(IProductRepository productRepo,
+            IRepository<Category> categoryRepo,
+            IRepository<ApplicationType> appTypeRepo,
+            ProductImageService productImageService)
         {
-            _dbContext = dbContext;
+            _productRepo = productRepo;
+            _categoryRepo = categoryRepo;
+            _appTypeRepo = appTypeRepo;
             _productImageService = productImageService;
         }
 
         [HttpGet]
         public IActionResult Index()
         {
-            var products = _dbContext.Products
-                .Include(p => p.Category)
-                .Include(p => p.ApplicationType);
+            var products = _productRepo.GetAllIncludeAll();
             return View(products);
         }
 
@@ -39,7 +44,7 @@ namespace RockyShop.Controllers
             if (id == null)
                 return View(viewModel);
 
-            viewModel.Product = _dbContext.Products.Find(id);
+            viewModel.Product = _productRepo.FirstOrDefaultIncludeAll(p => p.Id == id);
             if (viewModel.Product == null)
                 return NotFound();
             return View(viewModel);
@@ -77,14 +82,13 @@ namespace RockyShop.Controllers
                     return View(fromRequest);
 
                 fromRequest.Product.Image = newFileName;
-                _dbContext.Products.Add(fromRequest.Product);
+                _productRepo.Add(fromRequest.Product);
             }
             //Update
             else
             {
-                var toUpdate = _dbContext.Products
-                    .AsNoTracking()
-                    .FirstOrDefault(p => p.Id == fromRequest.Product.Id);
+                var toUpdate = _productRepo
+                    .FirstOrDefault(filter: p => p.Id == fromRequest.Product.Id, isTracking: false);
                 if (toUpdate == null)
                     return NotFound();
                 string oldFileName = toUpdate.Image;
@@ -95,21 +99,19 @@ namespace RockyShop.Controllers
                     fromRequest.Product.Image = newFileName;
                 }
 
-                _dbContext.Products.Update(fromRequest.Product);
+                _productRepo.Update(fromRequest.Product);
             }
 
-            _dbContext.SaveChanges();
+            _productRepo.SaveChanges();
             return RedirectToAction(nameof(Index));
         }
 
         [HttpGet]
         public IActionResult Delete(int? id)
         {
-            var toDelete = _dbContext.Products
-                .Include(p => p.Category)
-                .Include(p => p.ApplicationType)
-                .Where(p => p.Id == id)
-                .FirstOrDefault();
+            if (id == null)
+                return NotFound();
+            var toDelete = _productRepo.FirstOrDefaultIncludeAll(p => p.Id == id);
             if (toDelete == null)
                 return NotFound();
             return View(toDelete);
@@ -120,23 +122,25 @@ namespace RockyShop.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult DeletePost(int? id)
         {
-            var toDelete = _dbContext.Products.Find(id);
+            if (id == null)
+                return NotFound();
+            var toDelete = _productRepo.Find((int)id);
             if (toDelete == null)
-                return NotFound("Product with such id was not found");
+                return NotFound();
             _productImageService.DeleteImage(toDelete.Image);
-            _dbContext.Products.Remove(toDelete);
-            _dbContext.SaveChanges();
-            return RedirectToAction("Index");
+            _productRepo.Remove(toDelete);
+            _productRepo.SaveChanges();
+            return RedirectToAction(nameof(Index));
         }
 
         private void PopulateProductVM(ProductVM viewModel)
         {
-            viewModel.CategoryDropDown = _dbContext.Categories.Select(c => new SelectListItem()
+            viewModel.CategoryDropDown = _categoryRepo.GetAll().Select(c => new SelectListItem()
             {
                 Text = c.Name,
                 Value = c.Id.ToString()
             });
-            viewModel.AppTypeDropDown = _dbContext.ApplicationTypes.Select(a => new SelectListItem()
+            viewModel.AppTypeDropDown = _appTypeRepo.GetAll().Select(a => new SelectListItem()
             {
                 Text = a.Name,
                 Value = a.Id.ToString()
